requirements.txt
appdirs==1.4.4
asttokens==2.2.1
attrs==22.2.0
Automat==22.10.0
Babel==2.10.3
backcall==0.2.0
bcrypt==3.2.2
beautifulsoup4==4.11.2
beniget==0.4.1
blinker==1.5
Bottleneck==1.3.5
Brlapi==0.8.4
Brotli==1.0.9
certifi==2022.9.24
chardet==5.1.0
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
constantly==15.1.0
contourpy==1.0.7
cryptography==38.0.4
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.3.2
decorator==5.1.1
defusedxml==0.7.1
distro==1.8.0
distro-info==1.5+deb12u1
et-xmlfile==1.0.1
executing==1.2.0
fastjsonschema==2.16.3
fonttools==4.38.0
fs==2.4.16
gast==0.5.2
GDAL==3.6.2
greenlet==2.0.2
html5lib==1.1
httplib2==0.20.4
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
iniconfig==1.1.1
ipython==8.5.0
jdcal==1.0
jedi==0.18.2
Jinja2==3.1.2
jsonpointer==2.3
jsonschema==4.10.3
jupyter_core==4.12.0
kiwisolver==0.0.0
lazr.restfulclient==0.14.5
lazr.uri==1.0.6
llvmlite==0.39.1
louis==3.24.0
lxml==4.9.2
lz4==4.0.2+dfsg
Mako==1.2.4.dev0
MarkupSafe==2.1.2
matplotlib==3.6.3
matplotlib-inline==0.1.6
more-itertools==8.10.0
mpmath==0.0.0
msgpack==1.0.3
mutagen==1.46.0
nbformat==5.5.0
networkx==2.8.8
numba==0.56.4
numexpr==2.8.4
numpy==1.24.2
oauthlib==3.2.2
odfpy==1.4.2
olefile==0.46
openpyxl==3.0.9
OWSLib==0.27.2
packaging==23.0
pandas==1.5.3
parso==0.8.3
pbr==5.10.0
pem==19.1.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
plotly==5.4.0
pluggy==1.0.0+repack
ply==3.11
prompt-toolkit==3.0.36
psycopg2==2.9.5
ptyprocess==0.7.0
pure-eval==0.0.0
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.20.1
pycryptodomex==3.11.0
pycups==2.0.1
pycurl==7.45.2
pydot==1.4.2
pygame==2.6.1
Pygments==2.14.0
PyGObject==3.42.2
pygraphviz==1.7
PyHamcrest==2.0.3
PyJWT==2.6.0
pynvim==0.4.2
pyOpenSSL==23.0.0
pyparsing==3.0.9
pyproj==3.4.1
PyQt5==5.15.9
PyQt5-sip==12.11.1
pyrsistent==0.18.1
pyserial==3.5
PySimpleSOAP==1.16.2
pysmbc==1.0.23
pytest==7.2.1
python-apt==2.6.0
python-dateutil==2.8.2
python-debian==0.1.49
python-debianbts==4.0.1
pythran==0.11.0
pytz==2022.7.1
pyxattr==0.8.1
pyxdg==0.28
PyYAML==6.0
QScintilla==2.13.3
reportbug==12.0.0
requests==2.28.1
rfc3987==1.3.8
scipy==1.10.1
service-identity==18.1.0
six==1.16.0
soupsieve==2.3.2
stack-data==0.6.2
sympy==1.11.1
tables==3.7.0
tenacity==8.2.1
traitlets==5.5.0
Twisted==22.4.0
ufoLib2==0.14.0
uritemplate==4.1.1
urllib3==1.26.12
wadllib==1.3.6
wcwidth==0.2.5
webcolors==1.11.1
webencodings==0.5.1
websockets==10.4
xdg==5
yt-dlp==2023.3.4
zope.interface==5.5.2
